Remove black-screen for audio, add more README
FFMPEG command for the audio streaming no longer makes the video black as it appears to be too much of a bottleneck.

README now states that OBS and FFMPEG must match bitrate.
It took me 2:30 hours of debugging to discover this alongside with the `-vf geq=0:128:128` command for audio streaming being too much of a bottleneck.